Fix of #493: No way to turn-off Hardware acceleration in AuthenticationActivity WebView
Issue: By default WebView is always hardwareAccelerated due to which on
some devices it crashes intermittently because of bad GPU drivers. There
is a way to turn-off hardware acceleration by setting webview layer type
to 'LAYER_TYPE_SOFTWARE' but as webView used in AuthenticationActivity
and AuthenticationDialog is not directly accessible to caller so there
is no way caller could disable it.

Changes: Added a method in AuthenticationSettings i.e.
setHardwareAcceleration, which caller can use to disable WebView
hardware acceleration.