Implement the resiliency feature
1) AuthenticationContext
added property boolean mExtendedLifetimeEnabled, an explicit switch, which a developer can use to opt in the outage behavior. Set false as default

2) AuthenticationResult
added property boolean mIsExtendedLifeTimeToken, to indicate whether the token being returned is a token only valid in terms of extended lifetime. Set false as default. The value will be set to be true only if the stale access token is returned, that is
- [ the server is down (500/503/504) ]
- [ outage mode is on (AuthenticationContext.mExtendedLifetimeEnabled == true) ]
- [ stale access token returned is not expired in the ext_expires_in ]

added property Date mExtendedExpiresOn to indicate the date of the extended expire in for the AT. Set null as default

added the constructors of AuthenticationResult for new properties

3) TokenCacheItem
added property Date mExtendedExpiresOn. While returning the stale token, the mExpiresOn will be valued with mExtendedExpiresOn to minimize impact on client side code

4) Oauth2
implemented the retry logic
For every AcquireToken we should retry once before failing ONLY for 500/503/504. This behavior should be present irrespective of the ExtendedLifeTime flag being set or not making our library more resilient. We should retry once after interval of one second in case of failure.

If we retried once and get 500/503/504 response while outage mode is on, a valid stale access token will be returned. Otherwise, an Authentication exception will be thrown.

5) DefaultTokenCacheStore (token delete logic)
An access token will be deleted only if the token expires in terms both expires_in and ext_expires_in.